quantum_tic_tac_toe_test.py
import cirq
import numpy as np
import random

def test_create_quantum_board():
    board = create_quantum_board()
    assert len(board) == 9, "Board should have 9 qubits"
    for qubit in board:
        assert isinstance(qubit, cirq.Qubit), "Each element in board should be a Qubit"

def test_play_quantum_tic_tac_toe():
    class Player:
        def get_move(self, circuit, board, symbol):
            return cirq.X(board[random.randint(0, 8)])

    player1 = Player()
    player2 = Player()
    result = play_quantum_tic_tac_toe(player1, player2)
    assert result is None or result in ('X', 'O'), "Result should be None or 'X' or 'O'"

def test_play_quantum_tic_tac_toe_with_winner():
    class Player:
        def get_move(self, circuit, board, symbol):
            if symbol == 'X':
                return cirq.X(board[0])
            return cirq.X(board[8])

    player1 = Player()
    player2 = Player()
    result = play_quantum_tic_tac_toe(player1, player2)
    assert result == 'X', "Player 1 should win"

def test_play_quantum_tic_tac_toe_with_draw():
    class Player:
        def get_move(self, circuit, board, symbol):
            return cirq.X(board[random.randint(0, 8)])

    player1 = Player()
    player2 = Player()
    result = play_quantum_tic_tac_toe(player1, player2)
    assert result is None, "The game should end in a draw"
